iox-#1345 Rename platform to configuration
